+++ library/inspector/src/main/java/io/mehow/laboratory/inspector/SectionFragment.kt
@@ -23,7 +23,7 @@ internal class SectionFragment : Fragment(R.layout.io_mehow_laboratory_feature_g
 
   private lateinit var layoutManager: SmoothScrollingLinearLayoutManager
   private val featureAdapter = FeatureAdapter(object : FeatureAdapter.Listener {
-    override fun onSelectFeature(feature: Feature<*>) = inspectorViewModel.selectFeature(feature)
+    override fun onSelectOption(option: Feature<*>) = inspectorViewModel.selectFeature(option)
 
     override fun onGoToFeature(feature: Class<Feature<*>>) {
       lifecycleScope.launch {