+++ library/inspector/src/test/java/io/mehow/laboratory/inspector/InspectorViewModels.kt
@@ -45,3 +45,9 @@ internal fun InspectorViewModel(
 )
 
 internal fun InspectorViewModel.sectionFlow() = sectionFlow(InspectorViewModel.defaultSection)
+
+internal fun InspectorViewModel.supervisedFeaturesFlow(
+  sectionName: String = InspectorViewModel.defaultSection,
+) = sectionFlow(sectionName).map { uiModels ->
+  uiModels.flatMap { it.models }.map { it.option to it.supervisedFeatures }
+}